+++ src/main/groovy/com/ofg/uptodate/finder/NewVersionFinderFactory.groovy
@@ -1,9 +1,0 @@
-package com.ofg.uptodate.finder
-
-import com.ofg.uptodate.UptodatePluginExtension
-import com.ofg.uptodate.dependency.Dependency
-
-interface NewVersionFinderFactory {
-
-    NewVersionFinder create(UptodatePluginExtension uptodatePluginExtension, List<Dependency> dependencies)
-}