+++ src/main/groovy/io/fourfinanceit/uptodate/dependency/VersionPatternMatcher.groovy
@@ -1,0 +1,16 @@
+package io.fourfinanceit.uptodate.dependency
+
+class VersionPatternMatcher {
+
+    private final String version
+
+    VersionPatternMatcher(String version) {
+        this.version = version
+    }
+
+    boolean matchesNoneOf(List<String> patterns) {
+        return !patterns.any {
+            version.matches(it)
+        }
+    }
+}