+++ src/main/groovy/io/fourfinanceit/uptodate/finder/NewVersionFinderFactory.groovy
@@ -1,0 +1,9 @@
+package io.fourfinanceit.uptodate.finder
+
+import io.fourfinanceit.uptodate.UptodatePluginExtension
+import io.fourfinanceit.uptodate.dependency.Dependency
+
+interface NewVersionFinderFactory {
+
+    NewVersionFinder create(UptodatePluginExtension uptodatePluginExtension, List<Dependency> dependencies)
+}