+++ src/main/groovy/io/fourfinanceit/uptodate/finder/RepositorySettings.groovy
@@ -1,0 +1,11 @@
+package io.fourfinanceit.uptodate.finder
+
+import groovy.transform.Immutable
+import groovy.transform.PackageScope
+
+@Immutable
+class RepositorySettings {
+
+    String repoUrl
+    boolean ignoreRepo
+}