+++ src/main/groovy/io/fourfinanceit/uptodate/finder/http/HttpConnectionSettings.groovy
@@ -1,0 +1,12 @@
+package io.fourfinanceit.uptodate.finder.http
+
+import groovy.transform.Immutable
+
+@Immutable
+class HttpConnectionSettings {
+
+    String url
+    ProxySettings proxySettings
+    int poolSize
+    int timeout
+}