+++ src/main/groovy/io/fourfinanceit/uptodate/finder/http/ProxySettings.groovy
@@ -1,0 +1,13 @@
+package io.fourfinanceit.uptodate.finder.http
+
+import groovy.transform.Immutable
+import groovy.transform.PackageScope
+
+@Immutable
+@PackageScope
+class ProxySettings {
+
+    String hostname
+    int port
+    String scheme
+}